requirements.txt
asgiref==3.6.0
autopep8==2.0.1
bleach==6.0.0
Django==3.2.16
django-extensions==3.2.1
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==5.0
django-simple-captcha==0.5.17
django-summernote==0.8.20.0
django-taggit==3.1.0
django_debug_toolbar==3.8.1
Pillow==9.4.0
pycodestyle==2.10.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
webencodings==0.5.1
mysqlclient
